project/requirements.txt
pillow>=9.0.0
pdf2docx>=0.5.6
markdown>=3.4.0
weasyprint>=58.0
pdfkit>=1.0.0
PyQt6>=6.3.0
python-docx>=0.8.11
lxml>=4.9.0
pygments>=2.13.0
pytest>=7.0.0 